% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_rd_to_html.R
\name{convert_rd_to_html}
\alias{convert_rd_to_html}
\title{Convert Rd to HTML with Custom CSS}
\usage{
convert_rd_to_html(rd_file_name)
}
\arguments{
\item{rd_file}{A string specifying the name of the Rd file (with extension) located in the "man" directory.}

\item{output_dir}{A string specifying the directory where the HTML file will be saved. Defaults to "man/html-Rd".}

\item{css_file}{A string specifying the path to the CSS file that will be linked in the HTML header. Defaults to "man/html-Rd/styles.css".}
}
\value{
None. The function is used for its side effect of generating an HTML file from an Rd file.
}
\description{
Converts an Rd file to an HTML file and adds a custom CSS link to the HTML header.
}
\examples{
convert_rd_to_html("strsplit1.Rd")
convert_rd_to_html("strsplit1.Rd", output_dir = "man/html-Rd", css_file = "man/html-Rd/styles.css")

}
